feat: 채팅 말풍선 기본 배치 추가
