+++ frontend/src/pages/Room/containers/index.tsx
@@ -1,6 +1,5 @@
 export { default as CanvasContainer } from './CanvasContainer';
 export { default as ChatContainer } from './ChatContainer';
-export { default as ChatInputContainer } from './ChatInputContainer';
 export { default as GateContainer } from './GateContainer';
 export { default as SocketContainer } from './SocketContainer';
 export { default as UserListContainer } from './UserListContainer';
