+++ frontend/src/pages/Room/index.scss
@@ -24,6 +24,18 @@ canvas {
 
 .user-list {
   list-style: none;
+
+  li {
+    position: relative;
+  }
+
+  .chat-balloon {
+    position: absolute;
+    top: -20px;
+    padding: 2px 10px;
+    background-color: #fff;
+    border-radius: 6px;
+  }
 }
 
 // 채팅
